+++ NEWS.md
@@ -1,3 +1,11 @@
+# SciViews 1.5.0
+
+-   Topics are now implemented in `SciViews::R`. One can specify, for instance, `SciViews::R("model")` to load more packages related to modeling with regressions. Current topics ore 'infer', 'model', 'explore', 'ml', 'ts' and 'spatial'. It is possible to specify more than one topic.
+
+-   The new function `sciviews_packages_topics()` lists all packages required for each topic.
+
+-   `SciViews::R` has now a new argument `lang=` to specify the default language to use. Hence, `SciViews::R(lang = "fr")` will set French as de default language used by functions like `data.io::read()`, or for plot axes labels or title.
+
 # SciViews 1.4.0
 
 -   {tidyverse} is not loaded any more, but a selected series of tidyverse packages useful for `SciViews::R`.